requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pyaml==24.4.0
Pygments==2.17.2
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.1
